Tidy up util types
