+++ src/index.ts
@@ -7,5 +7,6 @@ export * from './parse';
 export * from './resolve';
 export * from './serialize';
 export * from './submit';
+export * from './utils';
 export * from './visitor';
 export * from './validate';
