+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+export * from './guards';
+export * from './is-uri';
+export * from './types';
