+++ src/utils/types.ts
@@ -1,0 +1,7 @@
+export type Href = string | URL;
+
+export interface Hyperlink {
+  href: Href;
+}
+
+export type Target = Href | Hyperlink;
